change paxe format (#6)
* change paxe format

* max length check fix

+++ trex-paxe/src/main/java/com/github/trex_paxos/paxe/Channel.java
@@ -1,8 +1,8 @@
 package com.github.trex_paxos.paxe;
 
-public record Channel(byte value) {
-    public static final byte CONSENSUS = 0;
-    public static final byte KEY_EXCHANGE = (byte)255;
+public record Channel(short value) {
+    public static final short CONSENSUS = 0;
+    public static final short KEY_EXCHANGE = 255;
     static final Channel CONSENSUS_CHANNEL = new Channel(CONSENSUS);
     static final Channel KEY_EXCHANGE_CHANNEL = new Channel(KEY_EXCHANGE);
 }
